Update and DleteCategory added

+++ src/controllers/categoryController.js
@@ -1,4 +1,6 @@
 import Category from "../models/Category/categoryModel.js";
+import Product from "../models/Product/productModel.js";
+import { updateChildrenPaths } from "../utility/updateChildrenPaths.js";
 import buildCategoryTree from "../utils/buildCategoryTree.js";
 // controllers/categoryController.js
 
@@ -67,16 +69,24 @@ export const getAllCategories = async (req, res) => {
 
 // To update a particular category
 export const updateCategory = async (req, res) => {
-  const { id } = req.params; // Get the category ID from the request parameters
-  const { name, description, path, level } = req.body; // Get the updated data from the request body
-
+  const { _id } = req.params; // Get the category ID from the request parameters
+  const { name, slug, parent, path, level } = req.body; // Get the updated data from the request body
   try {
+    //Avoid duplicate slugs
+    const existing = await Category.findOne({ slug, _id: { $ne: _id } });
+    if (existing) {
+      return res.status(400).json({ message: "Slug already in use.." });
+    }
+
     // Find the category by ID and update it
     const updatedCategory = await Category.findByIdAndUpdate(
-      id,
-      { name, description, path, level },
+      _id,
+      { name, slug, parent, path, level },
       { new: true }
     );
+
+    //If 'a parent category' is changed, update path and level of its children.
+    await updateChildrenPaths(updatedCategory);
 
     // If the category is not found, return a 404 error
     if (!updatedCategory) {
@@ -96,6 +106,23 @@ export const updateCategory = async (req, res) => {
 export const deleteCategory = async (req, res) => {
   const { _id } = req.params;
   try {
+    //Find the categories whose parent field equals the current category’s _id
+    const childCategories = await Category.find({ parent: _id });
+    //If a category has children, deletion of category should not be done.
+    if (childCategories.length > 0) {
+      return res
+        .status(400)
+        .json({ message: "Cannot delete a category which has sub-categories" });
+    }
+
+    //If a category has products, it should not be deleted.
+    const products = await Product.find({ categoryId: _id });
+    if (products.length > 0) {
+      return res
+        .status(400)
+        .json({ message: "Cannot delete a category which has products" });
+    }
+
     const deletedCategory = await Category.findByIdAndDelete(_id);
     if (!deletedCategory) {
       return res.status(404).json({ message: "Could not find category" });
